spm_image.m plot timecourse of SPM dataset and onsets
in SPM package, the spm_image.m is used to display the images of SPM dataset. I adapted this file by adding two more functions: 1) it can display the timecouse of each voxel of the fmri dataset; 2) it can display the onsets of each condition in your experiment. So these two functions can help you check your data. Importantly, this file doesn't ruin the function of the original spm_image.m file. So please feel free to use it